requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
Flask==3.1.3
gTTS==2.5.4
gunicorn==25.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
requests==2.33.0
urllib3==2.6.3
Werkzeug==3.1.7
